updated photo model to have favorites, description and added api controllers for tags, descriptions, favorites. Updated authentication for multi-user support

+++ server/src/controllers/photoController.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 import dotenv from 'dotenv';
-import { Photo } from '../models/photoModel';
+import { IPhoto, Photo } from '../models/photoModel';
 import { logInfo } from '../utils/logger'
 import * as db from '../config/db';
 import { GridFSBucket } from 'mongodb';
@@ -9,6 +9,7 @@ import fs from 'fs';
 import * as auth from '../utils/auth';
 import * as mdata from '../utils/getMetadata';
 import { content } from 'googleapis/build/src/apis/content';
+import { exit } from 'process';
 
 interface MulterRequest extends Request {
     file?: Express.Multer.File;
@@ -36,7 +37,7 @@ export const uploadPhoto = async (req: MulterRequest, res: Response) => {
             return res.status(403).json({ error: "Invalid or expired token" });
         }
     
-        const { filename, path: originalPath } = file;
+        const { path: originalPath } = file;
         logInfo(file);
         const contentType = file.mimetype || 'application/octet-stream';
         logInfo("Content type: ", contentType);
@@ -44,7 +45,7 @@ export const uploadPhoto = async (req: MulterRequest, res: Response) => {
             logInfo("Invalid file type");
             return res.status(400).json({ error: "Invalid file type" });
         }
-        const gridFSFileId = await db.uploadFileToGridFS(originalPath, filename);
+        const gridFSFileId = await db.uploadFileToGridFS(originalPath, file.originalname);
         if (!gridFSFileId || gridFSFileId instanceof Error) {
             logInfo("Error uploading file to GridFS", gridFSFileId);
             return res.status(500).json({ error: "Error uploading file to GridFS" });
@@ -57,9 +58,13 @@ export const uploadPhoto = async (req: MulterRequest, res: Response) => {
         }
 
         const [successUserId, userId] = await auth.extractPayload(authToken);
+        if (!successUserId) {
+            logInfo("Error extracting user ID from token");
+            return res.status(500).json({ error: "Error extracting user ID from token" });
+        }
         const newPhoto = new Photo({
             userId: userId,
-            filename: filename,
+            filename: file.originalname,
             gridFSFileId: gridFSFileId,
             contentType: contentType,
             dateTaken: metadata.DateTimeOriginal || undefined,
@@ -111,6 +116,12 @@ export const deletePhoto = async (req: Request, res: Response) => {
             return res.status(403).json({ error: "Invalid or expired token" });
         }
         
+        const successIsOwner = await auth.isPhotoOwner(authToken, id);
+        if (successIsOwner instanceof Error || !successIsOwner) {
+            logInfo("Error checking if user is photo owner");
+            return res.status(500).json({ error: successIsOwner });
+        }
+
         const [success, result] = await db.deletePhoto(id);
         if (success) {
             logInfo("Photo deleted successfully: ", id);
@@ -137,10 +148,20 @@ export const getPhotos = async (req: Request, res: Response) => {
             logInfo("Invalid or expired token");
             return res.status(403).json({ error: "Invalid or expired token" });
         }
-        const { userId, dateTaken, tags } = req.query;
+        const { userId, dateTaken, tags, isFavorite } = req.query;
         if (!userId) {
             logInfo("No user ID provided");
             return res.status(400).json({ error: "Please provide a user ID" });
+        }
+
+        const [successUserIdExtracted, userIdExtracted] = await auth.extractPayload(authToken);
+        if (!successUserIdExtracted) {
+            logInfo("Error extracting user ID from token");
+            return res.status(500).json({ error: "Error extracting user ID from token" });
+        }
+        if (userIdExtracted != userId) {
+            logInfo("Unauthorized user");
+            return res.status(403).json({ error: "Unauthorized user" });
         }
 
         const filters: any = { userId };
@@ -151,6 +172,13 @@ export const getPhotos = async (req: Request, res: Response) => {
         }
         if (tags) {
             filters.tags = { $in: (tags as string).split(',') };
+        }
+
+        if (isFavorite == 'true') {
+            filters.isFavorite = (isFavorite == 'true');
+        }
+        else if (isFavorite == 'false') {
+            filters.isFavorite = (isFavorite == 'false');
         }
 
         const photos = await Photo.find(filters);
@@ -181,11 +209,24 @@ export const downloadPhoto = async (req: Request, res: Response) => {
             logInfo("Invalid or expired token");
             return res.status(403).json({ error: "Invalid or expired token" });
         }
-
-        const photo = await Photo.findOne({ gridFSFileId: new mongoose.Types.ObjectId(gridFSFileId) });
+        
+        const result = await db.getPhotoIdfromGridFSId(gridFSFileId);
+        if (result instanceof Error) {
+            logInfo("Photo not found");
+            return res.status(404).json({ error: "Photo not found" });
+        }
+        
+        const [photoId, photo] = result as [mongoose.Types.ObjectId, IPhoto];
+        
         if (!photo) {
             logInfo("Photo not found");
             return res.status(404).json({ error: "Photo not found" });
+        }
+
+        const successIsOwner = await auth.isPhotoOwner(authToken, photoId.toString());
+        if (successIsOwner instanceof Error || !successIsOwner) {
+            logInfo("Error checking if user is photo owner");
+            return res.status(500).json({ error: successIsOwner });
         }
 
         const filestream = await db.downloadFileFromGridFS(gridFSFileId.toString());
@@ -241,6 +282,12 @@ export const renamePhoto = async (req: Request, res: Response) => {
             return res.status(403).json({ error: "Invalid or expired token" });
         }
 
+        const successIsOwner = await auth.isPhotoOwner(authToken, id);
+        if (successIsOwner instanceof Error || !successIsOwner) {
+            logInfo("Error checking if user is photo owner");
+            return res.status(500).json({ error: successIsOwner });
+        }
+
         const renameResult = await db.renamePhoto(id, newFilename);
         if (renameResult instanceof Error) {
             logInfo("Error renaming photo: ", renameResult);
@@ -254,3 +301,109 @@ export const renamePhoto = async (req: Request, res: Response) => {
         return res.status(500).json({ error: "Error renaming photo" });
     }
 }
+
+export const updateMetadata = async (req: Request, res: Response) => {
+    const authToken = req.headers['authorization'];
+    const { id } = req.params;
+    const { tags, description } = req.body;
+
+    if (!authToken || authToken == '' || authToken == null || authToken.trim() == '') {
+        logInfo("Missing Authentication Header");
+        return res.status(403).json({ error: "Missing Authentication Header" });
+    }
+    if (!id) {
+        logInfo("No photo ID provided");
+        return res.status(400).json({ error: "Please provide a photo ID" });
+    }
+    if (!tags && !description) {
+        logInfo("No new metadata provided");
+        return res.status(400).json({ error: "Please provide new metadata" });
+    }
+    try {
+        const updatedToken = await auth.verifyToken(authToken);
+        if (updatedToken instanceof Error) {
+            logInfo("Invalid or expired token");
+            return res.status(403).json({ error: "Invalid or expired token" });
+        }
+
+        const successIsOwner = await auth.isPhotoOwner(authToken, id);
+        if (successIsOwner instanceof Error || !successIsOwner) {
+            logInfo("Error checking if user is photo owner");
+            return res.status(500).json({ error: successIsOwner });
+        }
+
+        // Update tags
+        if (tags) {
+            const updatedTagsResult = await db.updateTags(id, tags);
+            if (updatedTagsResult instanceof Error) {
+                logInfo("Error updating tags: ", updatedTagsResult);
+                return res.status(500).json({ error: "Error updating tags" });
+            }
+            logInfo("Tags updated successfully: ", tags);
+        }
+
+        // Update description
+        if (description) {
+            const updatedDescriptionResult = await db.updateDescription(id, description.toString());
+            if (updatedDescriptionResult instanceof Error) {
+                logInfo("Error updating description: ", updatedDescriptionResult);
+                return res.status(500).json({ error: "Error updating description" });
+            }
+            logInfo("Description updated successfully: ", description);
+        }
+
+        // Return success response
+        logInfo("Metadata updated successfully: ", id, tags, description);
+        return res.status(200).json({ message: "Metadata updated successfully", id, tags, description });
+    } catch (error) {
+        logInfo("Error updating metadata: ", error);
+        return res.status(500).json({ error: "Error updating metadata" });
+    }
+}
+
+export const toggleFavorite = async (req: Request, res: Response) => {
+    const authToken = req.headers['authorization'];
+    const { id } = req.params;
+
+    if (!authToken || authToken == '' || authToken == null || authToken.trim() == '') {
+        logInfo("Missing Authentication Header");
+        return res.status(403).json({ error: "Missing Authentication Header" });
+    }
+    if (!id) {
+        logInfo("No photo ID provided");
+        return res.status(400).json({ error: "Please provide a photo ID" });
+    }
+    
+    try {
+        const updatedToken = await auth.verifyToken(authToken);
+        if (updatedToken instanceof Error) {
+            logInfo("Invalid or expired token");
+            return res.status(403).json({ error: "Invalid or expired token" });
+        }
+
+        const successIsOwner = await auth.isPhotoOwner(authToken, id);
+        if (successIsOwner instanceof Error || !successIsOwner) {
+            logInfo("Error checking if user is photo owner");
+            return res.status(500).json({ error: successIsOwner });
+        }
+
+        const photo = await Photo.findById(id);
+        if (!photo) {
+            logInfo("Photo not found");
+            return res.status(404).json({ error: "Photo not found" });
+        }
+
+        const updateSuccess = await db.updateIsFavorite(id);
+        if (updateSuccess instanceof Error) {
+            logInfo("Error toggling favorite status: ", updateSuccess);
+            return res.status(500).json({ error: "Error toggling favorite status" });
+        }
+        const isFavorite = await Photo.findById(id).select('isFavorite');
+        logInfo("Photo favorite status toggled successfully: ", id, isFavorite);
+        return res.status(200).json({ message: "Photo favorite status toggled successfully", id, isFavorite });
+    } catch (error) {
+        logInfo("Error toggling favorite status: ", error);
+        return res.status(500).json({ error: "Error toggling favorite status" });
+    }
+}
+
